add sweet alert

+++ src/app/views/pages/setting/setting.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { ToastrService } from 'ngx-toastr';
 import { SettingsService } from '../../../services/settings.service';
+import { SweetAlertService } from '../../../services/sweet-alert.service';
 import { CommonModule } from '@angular/common';
 
 @Component({
@@ -13,8 +14,12 @@ import { CommonModule } from '@angular/common';
 export class SettingComponent implements OnInit {
   settings: any[] = [];
   newSetting = { key: '', value: '' };
+  isUpdate: boolean = false;
+  settingId: any;
 
-  constructor(private settingsService: SettingsService, private toastr: ToastrService) {}
+  constructor(private settingsService: SettingsService,
+    private sweetAlert: SweetAlertService,
+    private toastr: ToastrService) {}
 
   ngOnInit() {
     this.fetchSettings();
@@ -26,26 +31,40 @@ export class SettingComponent implements OnInit {
     });
   }
 
-  addSetting() {
+  saveSetting() {
     if (!this.newSetting.key || !this.newSetting.value) return;
-
-    this.settingsService.addSetting(this.newSetting).then(() => {
-      this.toastr.success('Setting added successfully!');
-      this.newSetting = { key: '', value: '' };
-    });
+    if (this.isUpdate) {
+      this.settingsService.updateSetting(this.settingId, this.newSetting)
+      .then(() => {
+        this.toastr.success('Setting updated successfully!');
+        this.newSetting = { key: '', value: '' };
+        this.isUpdate = false;
+      });
+    } else {
+      this.settingsService.addSetting(this.newSetting).then(() => {
+        this.toastr.success('Setting added successfully!');
+        this.newSetting = { key: '', value: '' };
+      });
+    }
   }
 
-  updateSetting(settingId: string) {
-    const updatedData = { value: 'Updated Value' };
-
-    this.settingsService.updateSetting(settingId, updatedData).then(() => {
-      this.toastr.success('Setting updated successfully!');
-    });
+  showUpdateForm(settingId: string) {
+    this.isUpdate = true;
+    this.settingId = settingId;
+    this.settingsService.getSettingById(settingId).subscribe((data:any)=> {
+      this.newSetting = { key: data['key'], value: data['value']};
+    })
   }
 
   deleteSetting(settingId: string) {
-    this.settingsService.deleteSetting(settingId).then(() => {
-      this.toastr.warning('Setting deleted!');
+    this.sweetAlert.deleteConfirmation().then((result:any) => {
+      if (result.isConfirmed) {
+        this.settingsService.deleteSetting(settingId).then(() => {
+          this.toastr.success('Setting deleted successfully!');
+        });
+      }else{
+        this.toastr.warning('Deletion cancelled!');
+      }
     });
   }
 }
